logic_control/pyproject.toml
[project]
name = "logic_control"
version = "0.0.0"
authors = [{ name = "Your Name", email = "email@email.com" }]
description = "logic_control"
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"

dependencies = ["dora-rs >= 0.3.9"]

[dependency-groups]
dev = ["pytest >=8.1.1", "ruff >=0.9.1"]

[project.scripts]
logic_control = "logic_control.main:main"

[tool.ruff.lint]
extend-select = [
  "D",   # pydocstyle
  "UP"
]
